Documents/SEM6_3-2/SmartPlace_MINI_Project/SmartPlace_Mini_Project/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic
pydantic[email]
python-jose[cryptography]
passlib[bcrypt]
python-multipart
spacy
sentence-transformers
PyMuPDF
